move star stuff to folder
