+++ ChargerAstronomyEngine/Data/Star/AllStars.csv
