+++ ChargerAstronomyEngine/Data/Star/EquatorialStarMap.cs
